+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:54832/","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:64645/","mongoDBName":"jest"}
