+++ src/infrastructure/db/mongodb/log-repository/log.ts
@@ -1,4 +1,5 @@
-import { LogErrorRepository } from '@data/protocols/log-error-repository'
+
+import { LogErrorRepository } from '@data/protocols/db/log-error-repository'
 import { MongoHelper } from '../helpers/mongo-helper'
 
 export class LogMongoRepository implements LogErrorRepository {
