+++ src/main/decorators/log.spec.ts
@@ -1,8 +1,8 @@
 import { Controller, HttpRequest, httpResponse } from '@presentation/protocols'
 import { LogControllerDecorator } from './log'
 import { serverError, ok } from '@presentation/helpers/http/http-helper'
-import { LogErrorRepository } from '@data/protocols/log-error-repository'
 import { AccountModel } from '@domain/models/account'
+import { LogErrorRepository } from '@data/protocols/db/log-error-repository'
 
 const makeController = (): Controller => {
   class ControllerStub implements Controller {
