+++ src/main/decorators/log.ts
@@ -1,5 +1,5 @@
+import { LogErrorRepository } from '@data/protocols/db/log-error-repository'
 import { Controller, HttpRequest, httpResponse } from '@presentation/protocols'
-import { LogErrorRepository } from '@data/protocols/log-error-repository'
 
 export class LogControllerDecorator implements Controller {
   private readonly controller: Controller
